refactor OPMMirrorHandler to subclass TensorStoreHandler. Passes demo test.
